bump version for proper provides installation
If you already have some of these modules installed, Encode::Windows1252 will fail to be built properly. Not sure if its because of the recent rakudo path changes or the provides changes. In my case the module builds and claims to have been installed correctly, but HTTP::UserAgent failed to find it until I bumped the version.